KL-155/style: rename oauth to oauth2
