+++ src/main/java/taco/klkl/domain/oauth2/service/Oauth2KakaoLoginService.java
@@ -1,8 +1,8 @@
-package taco.klkl.domain.oauth.service;
+package taco.klkl.domain.oauth2.service;
 
 import org.springframework.stereotype.Service;
 
-import taco.klkl.domain.oauth.dto.request.KakaoUserInfoRequest;
+import taco.klkl.domain.oauth2.dto.request.KakaoUserInfoRequest;
 import taco.klkl.domain.user.dto.response.UserDetailResponse;
 
 @Service
