+++ src/main/java/taco/klkl/domain/oauth2/service/Oauth2KakaoLoginServiceImpl.java
@@ -1,4 +1,4 @@
-package taco.klkl.domain.oauth.service;
+package taco.klkl.domain.oauth2.service;
 
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
@@ -6,9 +6,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import taco.klkl.domain.oauth.dao.OauthRepository;
-import taco.klkl.domain.oauth.domain.Oauth;
-import taco.klkl.domain.oauth.dto.request.KakaoUserInfoRequest;
+import taco.klkl.domain.oauth2.dao.Oauth2Repository;
+import taco.klkl.domain.oauth2.domain.Oauth2;
+import taco.klkl.domain.oauth2.dto.request.KakaoUserInfoRequest;
 import taco.klkl.domain.user.domain.Gender;
 import taco.klkl.domain.user.domain.User;
 import taco.klkl.domain.user.dto.request.UserCreateRequest;
@@ -23,7 +23,7 @@ import taco.klkl.global.util.UserUtil;
 @RequiredArgsConstructor
 public class Oauth2KakaoLoginServiceImpl implements Oauth2KakaoLoginService {
 
-	private final OauthRepository oauthRepository;
+	private final Oauth2Repository oauth2Repository;
 	private final UserService userService;
 	private final UserUtil userUtil;
 
@@ -38,15 +38,15 @@ public class Oauth2KakaoLoginServiceImpl implements Oauth2KakaoLoginService {
 		Long oauthId = userInfoRequest.id();
 
 		// 이미 oauth로그인 기록이 있는 유저를 처리합니다.
-		if (oauthRepository.existsByOauthId(oauthId)) {
-			Oauth oauth = oauthRepository.findFirstByOauthId(oauthId);
-			User user = oauth.getUser();
+		if (oauth2Repository.existsByOauthId(oauthId)) {
+			Oauth2 oauth2 = oauth2Repository.findFirstByOauthId(oauthId);
+			User user = oauth2.getUser();
 			return UserDetailResponse.from(user);
 		}
 
 		User user = registerUser(userInfoRequest);
-		Oauth oauth = Oauth.of(user, userInfoRequest.id());
-		oauthRepository.save(oauth);
+		Oauth2 oauth2 = Oauth2.of(user, userInfoRequest.id());
+		oauth2Repository.save(oauth2);
 
 		return UserDetailResponse.from(user);
 	}
