+++ src/main/java/taco/klkl/domain/oauth2/service/Oauth2KakaoService.java
@@ -1,4 +1,4 @@
-package taco.klkl.domain.oauth.service;
+package taco.klkl.domain.oauth2.service;
 
 import org.springframework.stereotype.Service;
